Feat:separate views
