+++ .gitignore
@@ -1,3 +1,6 @@
+# config file for mysql id pass #
+config.ini
+
 # Django #
 *.log
 *.pot
